audio-training-seq/onetime-predict/requirements.txt
pyee
tensorflow == 2.13.0
tensorflow_io
numpy
scramjet-framework-py